fixes [ch13052].  The communications catch tests rely upon mbedtls., which has been moved to the `crypto` module. The ad-hoc makefile has been changed to a regular module. so that we can use other modules as dependencies.
