+++ .gitignore
@@ -53,3 +53,4 @@ nbproject/
 build/target/
 build/cores/
 hal/soft-ap/.idea/
+cmake-build-debug
