+++ communication/makefile
@@ -2,6 +2,9 @@
 MODULE=communication
 COMMUNICATION_MODULE_PATH=.
 
+# ensure the unit test is not built
+MODULE_BUILD_PATH=$(COMMUNICATION_MODULE_PATH)/src
+
 # Target this makefile is building.
 TARGET_TYPE = a
