+++ communication/tests/catch/import.mk
@@ -1,0 +1,7 @@
+COMMUNICATION_CATCH_TESTS_MODULE_NAME = communication_catch_tests
+COMMUNICATION_CATCH_TESTS_MODULE_PATH ?= $(PROJECT_ROOT)/communication/tests/catch
+
+COMMUNICATION_CATCH_TESTS_BUILD_PATH_EXT = $(BUILD_TARGET_PLATFORM)
+
+
+include $(call rwildcard,$(COMMUNICATION_CATCH_TESTS_MODULE_PATH)/,include.mk)
